+++ xmake/core/tool/builder.lua
@@ -137,9 +137,10 @@ function builder:_inherit_from_targetdeps(results, target, flagname)
         local dep = orderdeps[total + 1 - idx]
 
         -- is static or shared target library? link it
-        local depkind = dep:get("kind")
-        local targetkind = target:get("kind")
-        if depkind == "static" or depkind == "shared" then
+        local depkind      = dep:get("kind")
+        local targetkind   = target:get("kind")
+        local depconfig    = table.wrap(target:depconfig(dep:name()))
+        if (depkind == "static" or depkind == "shared") and (depconfig.inherit == nil or depconfig.inherit) then
             if flagname == "links" and (targetkind == "binary" or targetkind == "shared") then
 
                 -- add dependent link